Only initialize once on multiuser devices
On devices with multiple users (or work profiles), the boot receiver is
called for each user. But if the most recently signed-in user logs out
(or if work profile is turned off), the incall audio callback are not
dead (invalid), therefore incall audio will stop working.

To solve this, only initialize on the main system user.